# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  gamemanagement.CreateGameRequest:
    properties:
      allocated_time_minutes:
        minimum: 1
        type: integer
      away_team_id:
        type: string
      division_pool_id:
        type: string
      field_location_id:
        type: string
      first_pull_by:
        type: string
      game_round_id:
        type: string
      home_team_id:
        type: string
      metadata:
        additionalProperties: true
        type: object
      name:
        maxLength: 100
        type: string
      scheduled_time:
        type: string
      scorekeeper_id:
        type: string
    required:
    - allocated_time_minutes
    - away_team_id
    - division_pool_id
    - field_location_id
    - home_team_id
    - name
    - scheduled_time
    type: object
...
  gamemanagement.FieldSummaryDTO:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
  gamemanagement.GameDTO:
    properties:
      actual_end_time:
        type: string
      actual_start_time:
        type: string
      allocated_time_minutes:
        type: integer
      away_team:
        $ref: '#/definitions/gamemanagement.TeamSummaryDTO'
      away_team_score:
        type: integer
      created_at:
        type: string
      field_location:
        $ref: '#/definitions/gamemanagement.FieldSummaryDTO'
      first_pull_by:
        type: string
      game_round:
        $ref: '#/definitions/gamemanagement.GameRoundSummaryDTO'
      home_team:
        $ref: '#/definitions/gamemanagement.TeamSummaryDTO'
      home_team_score:
        type: integer
      id:
        type: string
      metadata:
        additionalProperties: true
        type: object
      name:
        type: string
      scheduled_time:
        type: string
      scorekeeper:
        $ref: '#/definitions/gamemanagement.UserSummaryDTO'
      status:
        type: string
      stoppage_time_seconds:
        type: integer
      updated_at:
        type: string
      version:
        type: integer
    type: object
  gamemanagement.GameEventDTO:
    properties:
      created_at:
        type: string
      description:
        type: string
      event_type:
        type: string
      id:
        type: string
      metadata:
        additionalProperties: true
        type: object
      minute:
        type: integer
      second:
        type: integer
    type: object
...
  gamemanagement.GameRoundSummaryDTO:
    properties:
      id:
        type: string
      name:
        type: string
      round_type:
        type: string
    type: object
  gamemanagement.GameTimelineDTO:
    properties:
      events:
        items:
          $ref: '#/definitions/gamemanagement.GameEventDTO'
        type: array
    type: object
  gamemanagement.RecordScoreRequest:
    properties:
      assists:
        minimum: 0
        type: integer
      blocks:
        minimum: 0
        type: integer
      goals:
        minimum: 0
        type: integer
      minute:
        type: integer
      player_id:
        type: string
      second:
        type: integer
      turns:
        minimum: 0
        type: integer
    required:
    - player_id
    type: object
  gamemanagement.RecordStoppageRequest:
    properties:
      duration_seconds:
        minimum: 1
        type: integer
      reason:
        maxLength: 255
        type: string
    required:
    - duration_seconds
    - reason
    type: object
  gamemanagement.ScoringDTO:
    properties:
      assists:
        type: integer
      blocks:
        type: integer
      created_at:
        type: string
      goals:
        type: integer
      id:
        type: string
      player_id:
        type: string
      player_name:
        type: string
      turns:
        type: integer
      updated_at:
        type: string
    type: object
...
  gamemanagement.StartGameRequest:
    properties:
      first_pull_by:
        type: string
    type: object
...
  gamemanagement.TeamSummaryDTO:
    properties:
      id:
        type: string
      logo_url:
        type: string
      name:
        type: string
    type: object
  gamemanagement.UpdateGameRequest:
    properties:
      allocated_time_minutes:
        minimum: 1
        type: integer
      first_pull_by:
        type: string
      metadata:
        additionalProperties: true
        type: object
      name:
        maxLength: 100
        type: string
      scheduled_time:
        type: string
      scorekeeper_id:
        type: string
    type: object
...
  gamemanagement.UserSummaryDTO:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
...
host: localhost:4000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API for DigiGameStats reimplementation.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: DigiGameStats API
  version: "1.0"
paths:
...
  /divisions/{id}/games:
    post:
      consumes:
      - application/json
      description: Create a new game schedule
      parameters:
      - description: Game Schedule Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/gamemanagement.CreateGameRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/gamemanagement.GameDTO'
        "400":
          description: bad request
          schema:
            type: string
        "409":
          description: field conflict
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Schedule Game
      tags:
      - games
...
  /games:
    get:
      description: List games with optional filters
      parameters:
      - description: Division Pool ID
        in: query
        name: division_pool_id
        type: string
      - description: Game Status
        in: query
        name: status
        type: string
      - description: Field ID
        in: query
        name: field_id
        type: string
      - description: Start Date (RFC3339)
        in: query
        name: start_date
        type: string
      - description: End Date (RFC3339)
        in: query
        name: end_date
        type: string
      - default: 50
        description: Limit results
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset for pagination
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/gamemanagement.GameDTO'
            type: array
        "400":
          description: bad request
          schema:
            type: string
      summary: List Games
      tags:
      - games
  /games/{id}:
    delete:
      description: Cancel a scheduled or in-progress game
      parameters:
      - description: Game ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: bad request
          schema:
            type: string
        "404":
          description: not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Cancel Game
      tags:
      - games
    get:
      description: Get game details by ID
      parameters:
      - description: Game ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gamemanagement.GameDTO'
        "404":
          description: not found
          schema:
            type: string
      summary: Get Game
      tags:
      - games
    put:
      consumes:
      - application/json
      description: Update game details
      parameters:
      - description: Game ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/gamemanagement.UpdateGameRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gamemanagement.GameDTO'
        "400":
          description: bad request
          schema:
            type: string
        "404":
          description: not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update Game
      tags:
      - games
  /games/{id}/end:
    post:
      description: Finalize game submission by scorekeeper
      parameters:
      - description: Game ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gamemanagement.GameDTO'
        "400":
          description: bad request
          schema:
            type: string
        "401":
          description: unauthorized
          schema:
            type: string
        "404":
          description: not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: End Game
      tags:
      - games
  /games/{id}/finish:
    post:
      description: Mark game timer as finished (scores can still be edited)
      parameters:
      - description: Game ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gamemanagement.GameDTO'
        "400":
          description: bad request
          schema:
            type: string
        "401":
          description: unauthorized
          schema:
            type: string
        "404":
          description: not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Finish Game
      tags:
      - games
  /games/{id}/scores:
    get:
      description: Get all player scores for a game
      parameters:
      - description: Game ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/gamemanagement.ScoringDTO'
            type: array
        "404":
          description: not found
          schema:
            type: string
      summary: Get Game Scores
      tags:
      - games
    post:
      consumes:
      - application/json
      description: Record player goals, assists, blocks, etc.
      parameters:
      - description: Game ID
        in: path
        name: id
        required: true
        type: string
      - description: Score Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/gamemanagement.RecordScoreRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gamemanagement.GameDTO'
        "400":
          description: bad request
          schema:
            type: string
        "401":
          description: unauthorized
          schema:
            type: string
        "404":
          description: not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Record Score
      tags:
      - games
...
  /games/{id}/start:
    post:
      consumes:
      - application/json
      description: Start game timer and set status to in_progress
      parameters:
      - description: Game ID
        in: path
        name: id
        required: true
        type: string
      - description: Start Game Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/gamemanagement.StartGameRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gamemanagement.GameDTO'
        "400":
          description: bad request
          schema:
            type: string
        "401":
          description: unauthorized
          schema:
            type: string
        "404":
          description: not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Start Game
      tags:
      - games
  /games/{id}/stoppages:
    post:
      consumes:
      - application/json
      description: Record game stoppage and extend game time
      parameters:
      - description: Game ID
        in: path
        name: id
        required: true
        type: string
      - description: Stoppage Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/gamemanagement.RecordStoppageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gamemanagement.GameDTO'
        "400":
          description: bad request
          schema:
            type: string
        "401":
          description: unauthorized
          schema:
            type: string
        "404":
          description: not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Record Stoppage
      tags:
      - games
  /games/{id}/stream:
    get:
      description: Get real-time game updates via Server-Sent Events
      parameters:
      - description: Game ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - text/event-stream
      responses:
        "200":
          description: event stream
          schema:
            type: string
      summary: Stream Game Updates
      tags:
      - games
  /games/{id}/timeline:
    get:
      description: Get complete game event timeline
      parameters:
      - description: Game ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gamemanagement.GameTimelineDTO'
        "404":
          description: not found
          schema:
            type: string
      summary: Get Game Timeline
      tags:
      - games
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"